+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## Unreleased
+
+* Upgraded libs
+* Removed indent functions. Fringe case.
+* Removed subStrBetween, subStrAfter, subStrUntil as they are available upstream now through alternative methods.
+* Added valueOf to String. Allows transformation of other types to string.
+* Added Boolean type. Added valueOf to boolean. Added tests for Boolean type.
+
 ## 1.1.1 - 2015-05-02
 
 * Fixed indentSize implementation.